refactor: move tabInputUri helper to src/common/utils
The helper is not scratchpad-specific, so keep it with the other shared
utils next to themeUtils.
